+++ Changelog.md
@@ -4,3 +4,4 @@
 - ComboBox for alignment input of PV 
 - Show EV chargers withouth smartmeter interface on dashboard
 - Update Strings in Screen Heatpump Info Screen.
+- Infoicon + Text "behavior on grid consumption" for charging mode "PV excess only"